Save backup of new logger WIP
